added "strict" option to validate function
Strict validation must occur before an object can be saved to the database.
Lenient (strict=false) validation is performed on objects that are already
in the database.  Soon, when strict=True it will test the parent elements
to make sure they are the same type of resource.  This shouldn't be done
when simply loading an object from the database and testing it for validity.
In particular, it would set off a chain reaction causing all objects it
depends on to be validated each time an object is loaded.