phrase choice lessons, as a temporary measure until flashcards are merged
This is basically copied directly from picture_choice, with modifications
to support both the prompt and the answers being textual phrases.

Since we plan to deprecate both these modules soon when flashcard support is
merged, the easiest plan at the moment was to just create a fork of
picture_choice that supports phrases instead.